Add functionality to buttons and keep score
The rock, paper, scissors buttons were given eventListeners that when
clicked, hold the user's choice, and randomly generate a choice for the
computer. A result then displays the outcome of the round, and each
time the button is clicked, it turns into a new round which is kept
track of by a round node.